// src/types/location.ts
// Vietnamese location DTOs used for province/district/ward dropdowns

export type ProvinceOption = {
  code: number;
  name: string;
  codename: string;
};

export type DistrictOption = {
  code: number;
  name: string;
  codename: string;
  division_type: string;
};

export type WardOption = {
  code: number;
  name: string;
  codename: string;
  division_type: string;
};